simplify + refactor
