+++ packages/server/src/vtl/lambdaRequestMappingTemplate.vtl
