+++ packages/server/src/vtl/lambdaResponseMappingTemplate.vtl
